legend.py: Fix crash due to broken fingerprint handler
According to the DETAILS file in GNUPG source code, the field 10 is special.
Changes in the commit e7f48e770c6856ca51f69 which focused on unifying the
handling of all versions/options broke the fingerprint handling as it did not
consider the different nature of --fingerprint. This commit reverses that change
and adds a note stating that fingerprint is different. It also fixes the UAT
handling as Field 10 is used to store the attribute subpacket count and size in
the UAT records.